refactor(2nd pass): fix P0 bugs, P1 improvements, P2 polish (20 files)
P0 — Bug fixes:
- MonitorEngine: eliminate inout+Task data race — webhook now runs
  synchronously on scan queue with DispatchSemaphore, results
  written back to event via HistoryStore.updateEvent()
- DirectoryScanner: visited set now shared across recursion levels
  (was creating new set per level, allowing infinite recursion)
- LevelSizeCalculator: relativePathString uses URL pathComponents
  instead of string replacement (was mismatching /foo/bar vs /foo/barbaz)

P1 — Important improvements:
- HistoryStore: add updateEvent() for webhook result writeback
- ByteFormatter: add UInt64 overload (eliminates Int64 overflow risk)
- WebhookManager: unified to Logger.shared (was os.Logger outlier)
- Logger: cachedRetentionDays protected by queue (thread safety)
- Logger: warn() deprecated in favor of warning()
- Logger: force unwrap in rotateOldLogs → guard let
- FolderAnalyzer: resolvingSymlinksInPath (was standardizingPath)
- LevelSizeCalculator: resolvingSymlinksInPath (same fix)
- LogViewer: DEBUG level support in regex + color (cyan)
- LogViewer: safe unwrap for regex match ranges
- LogViewer: efficient large file read via FileHandle (last 512KB)
- DateUtils: formatRelative handles future dates (clock skew)
- ConfigStore/StorageUtils: guard .first! with fallback

P2 — Polish:
- TemplateHelpers: removed unused sanitize() method
- All templates: direct UInt64 ByteFormatter calls (no Int64 cast)
- HistoryView: ForEach(event.levels) instead of ForEach(indices)
- SettingsView: cached WebhookManager instance for test sends
- DirectoryScanner: resolvingSymlinksInPath for proper loop detection

+++ BackupNotify/Notify/Templates/DiscordTemplate.swift
@@ -25,10 +25,10 @@ struct DiscordTemplate: NotificationTemplate {
         fields.append(makeField(name: "📁 Path", value: "`\(event.folderPath)`", inline: false))
         fields.append(makeField(name: "🕐 Created", value: DateUtils.displayString(from: event.createdAt), inline: true))
         fields.append(makeField(name: "🕐 Modified", value: DateUtils.displayString(from: event.modifiedAt), inline: true))
-        fields.append(makeField(name: "📊 Total Size", value: ByteFormatter.string(fromByteCount: Int64(event.totalSizeBytes)), inline: true))
+        fields.append(makeField(name: "📊 Total Size", value: ByteFormatter.string(fromByteCount: event.totalSizeBytes), inline: true))
         fields.append(makeField(name: "📄 Files", value: "\(event.fileCount)", inline: true))
         fields.append(makeField(name: "🎬 Videos", value: "\(event.videoCount)", inline: true))
-        fields.append(makeField(name: "🎬 Video Size", value: ByteFormatter.string(fromByteCount: Int64(event.videoSizeBytes)), inline: true))
+        fields.append(makeField(name: "🎬 Video Size", value: ByteFormatter.string(fromByteCount: event.videoSizeBytes), inline: true))
 
         if !event.levels.isEmpty {
             let levelsText = TemplateHelpers.formatLevelsText(event.levels, bullet: "•")
